Reorganized folder structure of examples
