+++ examples/tti/TTI_codegen.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 import numpy as np
 
 from devito.at_controller import AutoTuner
-from examples.tti_operators import *
+from examples.tti.tti_operators import *
 
 
 class TTI_cg:
