+++ examples/tti/tti_operators.py
